Привязка колледжа к площадке

+++ database/migrations/41_create_linking_the_college_to_the_site_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('linking_the_college_to_the_site', function (Blueprint $table) {
-            $table->integer('college_ID')->index();
-            $table->integer('site_ID')->index();
+            $table->unsignedBigInteger('college_ID')->index();
+            $table->unsignedBigInteger('site_ID')->index();
         });
     }
